Add bibliography to the manual
The custom citation style is copied from OSCAR and basically ensures
that the exact citation keys are used that we prescribe in the .bib file.

Citation keys are instead managed via `bibtool`. To this end, the helper
script `etc/format_bib` can be run from the root directory of this package to
reformat `docs/src/refs.bib` in a consistent way. If a custom citation
key is desired, one can add a `bibkey` field.